chore: move `joinPath` and `dirname` functions
